Piggyback on queries from peers.
Delay can now be set as a min/max pair. The next update will be
scheduled at a random point between those two.

A query is sent on init, and when the delay is hit, but the delay is
rescheduled when a peer beats us to it.

When a peer sends a query, we get to use that query's answers, and can
save a whole bunch of bandwidth and cpu by not sending a query
ourselves.