Remove invitation to connect
Comparison is used to prevent peers from connecting to eachother. When
this check fails, an additional mdns answer was broadcast to invite the
peer to connect.

This usually results in a flurry of spurious answers in even a small
swarm. In addition, it's rarely useful, because everyone responds to a
query anyway.